+++ src/schemas/contactFormSchema.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+
+export const contactFormSchema = z.object({
+  name: z.string().min(1, "Name is required"),
+  email: z.string().email("Invalid email"),
+  subject: z.string().min(1, "Subject is required"),
+  message: z.string().min(1, "Message is required"),
+});
+
+export type ContactFormData = z.infer<typeof contactFormSchema>;
